requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
configparser==7.1.0
dbus-python==1.3.2
gatt==0.2.7
h11==0.14.0
idna==3.10
paho-mqtt==2.1.0
pycairo==1.27.0
pydantic==2.10.1
pydantic_core==2.27.1
PyGObject==3.50.0
python-dotenv==1.0.1
requests==2.32.3
sniffio==1.3.1
starlette==0.41.3
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.32.1
